requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==4.2.19
django-allauth==65.4.1
django-crispy-forms==2.3
django-filter==25.1
pillow==11.1.0
sqlparse==0.5.3
